Uncertainty: Fix warning

+++ modules/Uncertainty/iAEntropy.cpp
@@ -54,7 +54,7 @@ void entropy(iAFilter* filter, QMap<QString, QVariant> const & parameters)
 	typedef itk::Image<PixelType, DIM> InputImageType;
 	typedef iAEntropyImageFilter<InputImageType, InputImageType> EntropyFilter;
 	auto entropyFilter = EntropyFilter::New();
-	for (int i = 0; i < filter->inputCount(); ++i)
+	for (size_t i = 0; i < filter->inputCount(); ++i)
 	{
 		entropyFilter->SetInput(i, dynamic_cast<InputImageType*>(filter->input(i)->itkImage()));
 	}
